+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 - Adds and default implementation of `verifySignatureAsync` to `PublicKey`.
 - Implement derivable abstracted account
+- Fix: Reverts experimental binary view functions, which caused `Buffer is undefined` errors in browsers.
 
 # 1.37.1 (2025-03-24)
 
@@ -38,9 +39,6 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 # 1.35.0 (2025-02-11)
 
 - Add `MultiEd25519Account` to support the legacy MultiEd25519 authentication scheme.
-- Add `experimental` section to SDK
-- Add experimental BCS support for view function outputs
-- Upgrade to aptos-client 1.0.0 / remove override for cross-spawn
 
 # 1.34.0 (2025-02-06)